refactor: extract unmute and completion-server into own files
Address review feedback:
- Split unmuteConversation from mute.ts into conversation/unmute.ts
- Extract completion-server handler from completion/index.ts into
  completion/server.ts
