+++ src/commands/conversation/index.ts
@@ -1,7 +1,8 @@
 import { Command } from 'commander'
 import { markConversationDone } from './done.js'
-import { muteConversation, unmuteConversation } from './mute.js'
+import { muteConversation } from './mute.js'
 import { replyToConversation } from './reply.js'
+import { unmuteConversation } from './unmute.js'
 import { showUnread } from './unread.js'
 import { viewConversation } from './view.js'
 import { findConversationWithUser } from './with.js'
